refactor: context 폴더 store로 이름변경
