+++ FE/issuetracker/src/components/Login/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import * as S from 'components/Login/styled.index';
-import { loginState } from 'context/loginState';
+import { loginState } from 'store/loginState';
 
 function Login() {
   const [login, setLogin] = useRecoilState(loginState);
